requirements.txt
alembic==1.5.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
